% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{annotate_junc_ref}
\alias{annotate_junc_ref}
\title{Annotate junctions using existing annotation}
\usage{
annotate_junc_ref(junc_metadata, gtf, use_strand = T)
}
\arguments{
\item{junc_metadata}{junction metadata output from \code{normalise_junc}. The
essential component is the junction co-ordinates in a GRanges format.}

\item{gtf}{either path to gtf or of class \code{ensemblGenome} loaded using
\code{\link{refGenome}}}
}
\value{
list. Raw and normalised counts including metdata detailing junction
  annotation.
}
\description{
\code{annotate_junc_ref} annotates junctions by whether their start and end
position precisely overlap with a known exon boundary. On this basis,
categorises junctions into "annotated", "novel_acceptor", "novel_donor",
"novel_combo", "exon_skip", "ambig_gene" and "none".
}
