% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{annotate_junc_ref}
\alias{annotate_junc_ref}
\alias{.get_ref_exons_annot}
\alias{.tidy_junc_annot}
\alias{.classify_junc}
\alias{.get_gr_for_start_end}
\alias{.merge_lists}
\title{Deprecated function: Annotate junctions using reference annotation}
\usage{
annotate_junc_ref(junc_metadata, gtf)

.get_ref_exons_annot(
  junc_metadata,
  ref_exons,
  ref_cols = c("strand", "gene_name", "gene_id", "transcript_id", "exon_id")
)

.tidy_junc_annot(
  junc_metadata,
  cols_to_merge = c("strand", "gene_name", "gene_id")
)

.classify_junc(junc_metadata, ref_junc)

.get_gr_for_start_end(gr)

.merge_lists(x, y)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\link[GenomicRanges:GRanges-class]{GRanges-class} format, the essential
component being the junction co-ordinates.}

\item{gtf}{either path to gtf or object of class \code{ensemblGenome} loaded
using \code{refGenome}.}
}
\value{
junction metadata as a \link[GenomicRanges:GRanges-class]{GRanges-class}
object with additional columns that detail overlapping
genes/transcripts/exons and junction categories.
}
\description{
\code{annotate_junc_ref} is deprecated, please use \code{junction_annot} instead.
annotates junctions by whether their start and end position precisely
overlaps with a known exon boundary. Using this information along with the
strand, junctions are categorised into "annotated", "novel_acceptor",
"novel_donor", "novel_combo", "novel_exon_skip", "ambig_gene" and "none".
}
\keyword{internal}
