% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.chr_filter}
\alias{.chr_filter}
\title{Filter a df for chromosomes to keep}
\usage{
.chr_filter(x, chrs)
}
\arguments{
\item{x}{df containing at minimum a column named "chr".}

\item{chrs}{chrs chromosomes to keep. If NULL, no filter is applied.}
}
\description{
\code{.chr_filter} will filter a df, keeping only chromosomes listed in the
chrs argument. It includes a couple of checks to notify users whether some or
all chrs are not found in their data. This aims to catch situations where
alternative nomenclature may be used - e.g. "M" vs "MT", or "chr1" vs "1".
}
