% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{.classify_junc}
\alias{.classify_junc}
\title{Classifys junctions}
\usage{
.classify_junc(junc_metadata, ref_junc)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}
}
\value{
junction metadata with additional columns informing whether that
  junction is present within annotation and it's category.
}
\description{
\code{.classify_junc} categories junctions into "annotated",
"novel_acceptor", "novel_donor", "novel_combo", "exon_skip", "ambig_gene" and
"none" using information from annotation and strand. Adds two additional
columns
}
