% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{.convert_hits_to_annot_ref}
\alias{.convert_hits_to_annot_ref}
\title{Converts overlapping hits into annotation}
\usage{
.convert_hits_to_annot_ref(
  junc_metadata,
  junc_exon_hits,
  ref_exons,
  ref_cols,
  col_suffix
)
}
\arguments{
\item{junc_exon_hits}{hits.}

\item{ref_exons}{gr.}

\item{ref_cols}{chr.}

\item{col_suffix}{chr.}
}
\value{
gr. Junction
}
\description{
\code{.convert_hits_to_annot_ref}
}
