% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_junc.R
\name{.get_count_filter}
\alias{.get_count_filter}
\title{Filter junctions}
\usage{
.get_count_filter(raw_count, count_thresh, n_samp)
}
\arguments{
\item{raw_count}{dataframe with columns as samples, rows as junctions, cells
as the raw count.}

\item{count_thresh}{the number of counts below which a junction will be
filtered out.}

\item{n_samp}{the number of samples that have to express the junction above
the \code{count_thresh}.}
}
\value{
logical vector with TRUE meaning the junction has passed the count
  filter.
}
\description{
\code{get_junc_count_filter} obtains a logical vector with TRUE marking the
junctions that are above the count filter specified.
}
