% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_gr_for_start_end}
\alias{.get_gr_for_start_end}
\title{Splits a \code{GRanges} object by it's start and end.}
\usage{
.get_gr_for_start_end(gr)
}
\arguments{
\item{gr}{Any \code{GRanges} object.}
}
\value{
list of 2 grs, each with 1 range corresponding to every range in the
  input. One contains start positions, the other ends.
}
\description{
\code{.get_gr_for_start_end} takes a \code{GRanges} object and generates 2,
one containing only the start co-ordinate and the other, the end.
}
