% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_junc.R
\name{.get_junc_cluster}
\alias{.get_junc_cluster}
\title{Obtain junction clusters}
\usage{
.get_junc_cluster(junc_metadata)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}
}
\value{
junction metadata with cluster definitions.
}
\description{
\code{.get_junc_cluster} defines a cluster for each junction. Each junction
is used as a seed to find other junctions that share it's acceptor or donor
position. Each cluster is comprised by it's seed junction along with any
junctions matching with a acceptor/donor position.
}
