% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{.get_ref_exons_annot}
\alias{.get_ref_exons_annot}
\title{Extracts annotation from the reference gtf}
\usage{
.get_ref_exons_annot(
  junc_metadata,
  ref_exons,
  ref_cols = c("strand", "gene_name", "gene_id", "transcript_id", "exon_id")
)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}

\item{ref_exons}{annotated exons imported from the gtf.}

\item{ref_cols}{column names matching the annotation columns from the
reference.}

\item{junc_start_end}{start and end of the junctions returned from
\code{\link{.get_gr_for_start_end}}.}

\item{ref_exons_start_end}{start and end of the exons returned from
\code{\link{.get_gr_for_start_end}}.}
}
\value{
junction metadata with annotation.
}
\description{
\code{.convert_hits_to_annot_ref} will find the overlap of each junction with
the annotated exons. Then, for each corresponding hit, annotates each
junction with the strand/exon/transcript/gene from the reference annotation.
}
