% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_load.R
\name{.junc_merge}
\alias{.junc_merge}
\title{Merge two junction datasets together}
\usage{
.junc_merge(junc_df_all, junc_df, i)
}
\description{
\code{.junc_merge} will merge two sets of junction data together. It uses a
full_join so will keep all rows from both datasets. It will also ensure
ambiguous strands ("*") are allowed to match with forward ("+") and reverse
("-") strands.
}
