% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_load.R
\name{.load_STAR}
\alias{.load_STAR}
\title{Load raw junction data}
\usage{
.load_STAR(junc_path)
}
\description{
\code{.load_STAR} will load raw junction data that is outputted from STAR
(SJ.out) into R. This will format the junction data, retaining only chr,
start, end, strand and count (uniq_map_read_count) columns.
}
