% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{.merge_lists}
\alias{.merge_lists}
\title{Merges two Characterlists into 1 with elementwise concatenation of the
vectors inside each list}
\usage{
.merge_lists(x, y)
}
\arguments{
\item{x}{list. Elements contain chr vectors.}

\item{y}{list. Elements contain chr vectors.}
}
\value{
list. Elements contain chr vectors concatenated between x and y
}
\description{
\code{.merge_lists}
}
