% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_junc.R
\name{.normalise_count}
\alias{.normalise_count}
\title{Normalise junction counts}
\usage{
.normalise_count(junc_metadata, raw_count)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}
}
\value{
dataframe matching the \code{raw_count} with columns as samples, rows
  as junctions with elements replaced with the normalised counts.
}
\description{
\code{.normalise_count} normalises the junction counts for each sample by
dividing the counts for each junction by the total number of counts of it's
associated cluster.
}
