% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasper-package.R
\name{.onLoad}
\alias{.onLoad}
\title{Load sklearn upon initialisation of dasper}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{placeholder}

\item{pkgname}{placeholder}
}
\description{
\code{.onLoad} uses superassignment to update global reference to sklearn.
Delay load means allows successfull load without sklearn installed. This
allows users to set python environment using reticulate::use_virtualenv().
}
\seealso{
https://rstudio.github.io/reticulate/articles/package.html
}
