% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{.tidy_junc_annot}
\alias{.tidy_junc_annot}
\title{Tidy junction annotation}
\usage{
.tidy_junc_annot(
  junc_metadata,
  cols_to_merge = c("strand", "gene_name", "gene_id")
)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}

\item{cols_to_merge}{}
}
\value{

}
\description{
\code{.tidy_junc_annot} merges the gene and strand details from the start and
end into one column per junction. Then, combines strand information from the
original rna-seq based and that from overlapping annotation.
}
