% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_juncs}
\alias{example_juncs}
\title{Random set of junctions from two example samples}
\format{A data frame with 19733 rows and 5 variables: \describe{
  \item{chr}{chromsome}
  \item{start}{first position of intron}
  \item{end}{last position of intron}
  \item{strand}{strand of intron}
  \item{eg1}{raw junction counts for sample "eg1"}
  \item{eg2}{raw junction counts for sample "eg1"}
  }}
\source{
generated using data-raw/example_juncs.R
}
\usage{
example_juncs
}
\description{
A dataset containing the example junction data for 2 samples outputted from
\code{\link{merge_junc}} used for vignettes and testing.
}
\keyword{datasets}
