% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_cov_norm.R
\name{junction_cov_norm}
\alias{junction_cov_norm}
\title{For each junction, obtain the normalised coverage across associated
exonic/intronic regions}
\usage{
junction_cov_norm(
  junctions,
  ref,
  unannot_width = 20,
  cov_paths_case,
  cov_paths_control,
  cov_chr_control = NULL
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{ref}{either path to gtf/gff3 or object of class \code{TxDb} imported using
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}

\item{unannot_width}{integer scalar determining the width of the region to
obtain coverage from when the end of of a junction does not overlap an
existing exon.}

\item{cov_paths_case}{paths to the BigWig/BAM files containing the coverage of
your case samples. Must be the same length and order to the rows of the
assays in \code{junctions}.}

\item{cov_paths_control}{paths to the BigWig/BAM files}

\item{cov_chr_control}{either "chr" or "no_chr", indicating the chromosome
format of control coverage data. To be used if you know the chromosome
format of the control BigWig/BAM files is different to that of your
junctions.}
}
\value{
list containing sublists, one for cases and the other controls. Each
sublist contains 3 matrices, corresponding the coverage for each sample
across the 2 exons and intron associated with every junction.
}
\description{
\code{junction_cov_norm} obtains regions of interest for each junction where
coverage disruptions would be expected. If ends of junctions are annotated,
it will use the overlapping exon definitions, picking the shortest exon when
multiple overlap one end. If unannotated, will use a user-defined width
(i.e. 20bp). To compare between samples (case/controls) coverage is
normalised to a set region. By default, the boundaries of each gene
associated to a junction are used as the region to normalise to.
}
\examples{

\dontrun{
# leave this as not run for now to save time for R CMD check
ref <- "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
ref <- GenomicFeatures::makeTxDbFromGFF(ref)
cov_paths_case <- list.files("/data/RNA_seq_diag/mito/bw/", full.names = T)[1:2]
cov_paths_control <- list.files("/data/recount/GTEx_SRP012682/gtex_bigWigs/all_gtex_tissues_raw_bigWigs/", full.names = T)[1:2]
cov <- junction_cov_norm(
    junctions_annot_example,
    ref,
    unannot_width = 20,
    cov_paths_case,
    cov_paths_control,
    cov_chr_control = "chr"
)
junctions
}

}
