% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_process.R
\name{junction_process}
\alias{junction_process}
\title{Wrapper function for processing junctions}
\usage{
junction_process(
  junctions,
  ref,
  count_thresh = c(raw = 5),
  n_samp = c(raw = 1),
  width_range = NULL,
  types = NULL,
  regions = NULL,
  score_func = .zscore,
  ...
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{ref}{either path to gtf/gff3 or object of class \code{TxDb} imported using
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}

\item{count_thresh}{named vector with names matching the names of the
\code{\link[SummarizedExperiment]{assays}} in \code{junctions}. Values
denote the number of counts below which a junction will be filtered out.}

\item{n_samp}{named vector with names matching the names of the
\code{\link[SummarizedExperiment]{assays}} in \code{junctions}. Values
denotes number of samples that have to express the junction above the
\code{count_thresh} in order for that junction to not be filtered.}

\item{width_range}{numeric vector of length 2. The first element denoting the
lower limit of junction width and the second the upper limit. Junctions
with widths outside this range will be filtered out.}

\item{types}{any junctions matching these types, derived form
\code{\link{junction_annot}} e.g. "unannotated" will be filtered out.}

\item{regions}{any junctions overlapping this set of regions (in a
\code{\link[GenomicRanges]{GRanges}} format) will be filtered out.}

\item{score_func}{function to score junctions by abnromality. By default,
will use a z-score. But this can be switched dependent on the format of the
user's junction data. Function must take as input an x and y argument. x
being the case junctions and y being the controls. This must return a
numeric vector equal to the length of x with each element corresponding to
a abnormality of a junction in each case sample.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object containing filtered, annotated, normalised junction data with scores.
}
\description{
\code{junction_process} wraps all "junction_" prefixed functions in
\code{dasper} except \link{junction_load}. This is designed to simplify
processing of the junction data for those familiar or uninterested with the
intermediates.
}
\examples{

ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
junctions_processed <-
    junction_process(
        junctions_example,
        ref,
        count_thresh = c("raw" = 5),
        n_samp = c("raw" = 1),
        width_range = c(25, 1000000),
        types = c("ambig_gene", "unannotated"),
    )
junctions_processed
}
