% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_score.R
\name{junction_score}
\alias{junction_score}
\title{Score patient junctions by their abnormality}
\usage{
junction_score(junctions, score_func = .zscore, ...)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{score_func}{function to score junctions by their abnormality. By default,
will use a z-score but can be switched to a user-defined function. This
function must take as input an \code{x} and \code{y} argument, containing case and
control counts respectively. This must return a numeric vector equal to the
length of \code{x} with elements corresponding to a abnormality of each junction.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
junctions as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object filtered for only case samples with an additional \code{assay}
containing junction abnormality scores.
}
\description{
\code{junction_score} will use the counts contained within the "norm"
\link[SummarizedExperiment:SummarizedExperiment-class]{assay} to calculate a
deviation of each patient junction from the expected distribution of control
junction counts. The function used to calculate this abnormality score can be
user-inputted or left as the default z-score. Junctions will also be labelled
based on whether they are up-regulated (+1) or down-regulated (-1) with
respect to controls junction and this information is stored in the
\link[SummarizedExperiment:SummarizedExperiment-class]{assay} "direction" for use
in \link{outlier_aggregate}.
}
\examples{

if (!exists("junctions_normed")) {
    junctions_normed <- junction_norm(junctions_example)
}
junctions_scored <- junction_score(junctions_normed)
junctions_scored
}
\seealso{
Junction processing functions
\code{\link{junction_annot}()},
\code{\link{junction_filter}()},
\code{\link{junction_load}()},
\code{\link{junction_norm}()},
\code{\link{junction_process}()}
}
\concept{junction}
