% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junctions_annot_example}
\alias{junctions_annot_example}
\title{Set of example annotated junctions}
\format{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object from \code{\link{SummarizedExperiment}} detailing the counts,
co-ordinates of junctions lying on chromosome 21/22 for 2 example samples and
3 controls: \describe{ \item{assays}{matrix with counts for junctions (rows)
and 5 samples (cols)} \item{colData}{example sample metadata}
\item{rowRanges}{\code{\link[GenomicRanges]{GRanges}} object describing the
co-ordinates and strand of each junction} \item{rowData}{annotation columns
detailing the overlapping exons/transcripts/genes of each junction and the
junction category}}
}
\source{
generated using data-raw/junctions_annot_example.R
}
\usage{
junctions_annot_example
}
\description{
The identical set of junctions as from \link{junctions_example} with
additional \code{rowData} giving the annotation of each junction, generated
by \link{junction_annot}.
}
\keyword{datasets}
