% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_junc.R
\name{normalise_junc}
\alias{normalise_junc}
\title{Normalise junction counts}
\usage{
normalise_junc(junc_metadata, raw_count)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format, the essential component being
the junction co-ordinates.}
}
\value{
list containing metadata in a \code{\link[GenomicRanges]{GRanges}}
  format and dataframes detailing raw and normalised counts.
}
\description{
\code{normalise_junc} normalises the junction counts for each sample. First,
each junction is clustered by finding all other junctions that share with it
an acceptor or donor position. Then, a proportion-spliced-in is calculated
for each junction by dividing the raw junction count by the total number of
counts in it's associated cluster.
}
