% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sashimi.R
\name{plot_sashimi}
\alias{plot_sashimi}
\title{Visualise RNA-seq data in a the form of a sashimi plot}
\usage{
plot_sashimi(
  junctions,
  ref,
  gene_tx_id,
  case_id = NULL,
  control_agg_func = NULL,
  region = NULL,
  annot_colour = c(ggpubr::get_palette("jco", 1), ggpubr::get_palette("npg", 7)[c(1, 3,
    2, 5, 6)], ggpubr::get_palette("jco", 6)[c(3)]),
  digits = 2,
  count_label = FALSE
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{ref}{either path to gtf/gff3 or object of class \link[GenomicFeatures:TxDb-class]{TxDb-class}.}

\item{gene_tx_id}{gene name as ensembl id, ensembl transcript id or gene
name.}

\item{case_id}{list of one element. This must be a character vector
containing the}

\item{control_agg_func}{list of one element. This must be a character vector
containing the}

\item{region}{a \link[GenomicRanges:GRanges-class]{GenomicRanges} of length 1
that is used to filter the exons/junctions. Only those that overlap this
region are plotted.}

\item{annot_colour}{character vector of colours for junction types. One value
must be supplied labelling each of the 7 possible types.}

\item{digits}{used in \code{round}, specifying the number of digits to round the
junction counts to for visualisation purposes.}

\item{count_label}{logical value specifying whether to add label the count of
each junction.}
}
\value{
\code{ggplot} displaying the splicing (and coverage) surrounding the
transcript/region of interest.
}
\description{
\code{plot_sashimi} plots the splicing events and coverage over a specific
genes/transcript and/or genomic region of interest. The plotting are built
from \code{ggplot2} functions.
}
\concept{plotting}
