% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ref_load}
\alias{ref_load}
\title{Load reference annotation into a TxDb format}
\usage{
ref_load(ref, organism = NA)
}
\arguments{
\item{ref}{either path to gtf/gff3 or object of class
\link[GenomicFeatures:TxDb-class]{TxDb-class}.}

\item{organism}{
    What is the Genus and species of this organism. Please use proper
    scientific nomenclature for example: "Homo sapiens" or "Canis familiaris"
    and not "human" or "my fuzzy buddy". If properly written, this information
    may be used by the software to help you out later.
  }
}
\value{
a \link[GenomicFeatures:TxDb-class]{TxDb-class} object.
}
\description{
\verb{ref_load`` loads reference annotation using [makeTxDbFromGFF][GenomicFeatures::makeTxDbFromGFF] if a character or leaves }ref` unchanged if already a \link[GenomicFeatures:TxDb-class]{TxDb-class}. If you would
}
