% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{test_annotate_junc_ref}
\alias{test_annotate_junc_ref}
\title{Tests annotate_junc_ref for whether N juncs are matching the correct exons
and also whether the annotation for acceptor donor fits expections}
\usage{
test_annotate_junc_ref(junc_metadata_w_ref, gtf, n_junc_to_test = 100)
}
\arguments{
\item{n_junc_to_test}{}
}
\value{

}
\description{
\code{test_annotate_junc_ref}
}
