% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_junc.R
\name{annotate_junc_ref}
\alias{annotate_junc_ref}
\title{Annotate junctions using reference annotation}
\usage{
annotate_junc_ref(junc_metadata, gtf)
}
\arguments{
\item{junc_metadata}{junction metadata in a
\code{\link[GenomicRanges]{GRanges}} format. The essential component is the
junction co-ordinates. Other metadata columns will be returned unchanged.}

\item{gtf}{either path to gtf or object of class \code{ensemblGenome} loaded
using \code{\link{refGenome}}.}
}
\value{
junction metadata with additional columns that detail overlapping
  genes/transcripts/exons and junction categories.
}
\description{
\code{annotate_junc_ref} annotates junctions by whether their start and end
position precisely overlap with a known exon boundary. On this basis, then
categorises junctions into "annotated", "novel_acceptor", "novel_donor",
"novel_combo", "novel_exon_skip", "ambig_gene" and "none".
}
