% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_score.R
\name{coverage_score}
\alias{coverage_score}
\title{Score coverage by their abnormality}
\usage{
coverage_score(junctions, coverage, score_func = .zscore, ...)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{coverage}{list containing normalised coverage data that is outputted
from \link{coverage_norm}.}

\item{score_func}{function to score junctions by their abnormality. By default,
will use a z-score but can be switched to a user-defined function. This
function must take as input an \code{x} and \code{y} argument, containing case and
control counts respectively. This must return a numeric vector equal to the
length of \code{x} with elements corresponding to a abnormality of each junction.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
junctions as
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
object with additional \code{assays} named "coverage_region" and
"coverage_score". "coverage_region" labels the region of greatest
disruption (1 = exon_start, 2 = exon_end, 3 = intron) and "coverage_score"
contains the abnormality scores of the region with the greatest disruption.
}
\description{
\code{coverage_score} will score disruptions in the coverage across the
intronic/exonic regions associated with each junction. This abnormality score
generated by \code{score_func} operates by calculating the deviation of the
coverage in patients to a coverage across the same regions in controls. Then,
for each junction it obtains the score of the region with the greatest
disruption.
}
\examples{

# use Genomic state to load txdb (GENCODE v31)
ref <- GenomicState::GenomicStateHub(version = "31", genome = "hg38", filetype = "TxDb")[[1]]

junctions_processed <- junction_process(
    junctions_example,
    ref,
    count_thresh = c("raw" = 5),
    n_samp = c("raw" = 1),
    types = c("ambig_gene", "unannotated"),
)

# obtain path to example bw on recount2
url <- recount::download_study(
    project = "SRP012682",
    type = "samples",
    download = FALSE
)

bw_path <- dasper:::.file_cache(url[1])

coverage_normed <- coverage_norm(
    junctions_processed,
    ref,
    unannot_width = 20,
    coverage_paths_case = rep(bw_path, 2),
    coverage_paths_control = rep(bw_path, 2),
    coverage_chr_control = "chr"
)

junctions_cov_scored <- coverage_score(junctions_processed, coverage_normed)
}
\seealso{
Coverage processing functions
\code{\link{coverage_norm}()},
\code{\link{coverage_process}()}
}
\concept{coverage}
