% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_score.R
\name{coverage_score}
\alias{coverage_score}
\title{Score coverage disruptions}
\usage{
coverage_score(junctions, coverage, score_func = .zscore, ...)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{coverage}{list containing normalised coverage data that is outputted from
\link{coverage_norm}.}

\item{score_func}{function to score junctions by abnromality. By default,
will use a z-score. But this can be switched dependent on the format of the
user's junction data. Function must take as input an x and y argument. x
being the case junctions and y being the controls. This must return a
numeric vector equal to the length of x with each element corresponding to
a abnormality of a junction in each case sample.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
junctions as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object with
additional assays named \code{coverage_region} and \code{coverage_score}.
\code{coverage_region} labels the region of greatest disruption (1 = exon_start,
2 = exon_end, 3 = intron) and \code{coverage_score} contains the scores of the
region with the greatest disruption.
}
\description{
\code{coverage_score} will score disruptions in the coverage across the
intronic/exonic regions associated with each junction. This abnormality score
is generated by \code{score_func} and operates by comparing coverage
originating from patients to a set of controls. Then, for each junction it
obtains the score of the region with the highest absolute score/greatest
disruption.
}
\examples{

\dontrun{
# leave this as not run for now to save time for R CMD check
ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
ref <-
    GenomicFeatures::makeTxDbFromGFF(ref)
junctions <-
    junction_norm(junctions_annot_example)
junctions <-
    junction_score(junctions)
coverage_paths_case <-
    list.files("/data/RNA_seq_diag/mito/bw/", full.names = T)[1:2]
coverage_paths_control <-
    list.files("/data/recount/GTEx_SRP012682/gtex_bigWigs/all_gtex_tissues_raw_bigWigs/",
        full.names = T
    )[1:2]
coverage <- coverage_norm(
    junctions,
    ref,
    unannot_width = 20,
    coverage_paths_case,
    coverage_paths_control,
    coverage_chr_control = "chr"
)

junctions
}

}
