% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.merge_lists}
\alias{.merge_lists}
\title{Merges two Characterlists into one with element-wise concatenation of the
vectors inside each list}
\usage{
.merge_lists(x, y)
}
\arguments{
\item{x}{CharacterList}

\item{y}{CharacterList}
}
\value{

}
\description{
\code{.merge_lists}
}
