% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_juncs}
\alias{example_juncs}
\title{Random set of example junctions from two samples}
\format{
RangedSummarizedExperiment from the
\code{\link{SummarizedExperiment}} detailing the counts, co-ordinates of
junctions for 2 example samples:
\describe{
\item{assays}{matrix with counts for ~20,000 junctions and 2 samples}
\item{colData}{example sample metadata}
\item{rowRanges}{GRanges object describing the co-ordinates and strand of each junction}
}
}
\source{
generated using data-raw/example_juncs.R
}
\usage{
example_juncs
}
\description{
A dataset containing the example junction data for 2 samples outputted from
\code{\link{junc_load}}.
}
\keyword{datasets}
