% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_annot.R
\name{junction_annot}
\alias{junction_annot}
\title{Annotate junctions using reference annotation}
\usage{
junction_annot(junctions, ref)
}
\arguments{
\item{junctions}{junction data as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, the
essential component being the junction co-ordinates as part of the
\code{rowRanges}. This can either be generated by \code{\link{junction_load}} or
independently constructed by the user.}

\item{ref}{either path to gtf/gff3 or object of class \code{TxDb} imported using
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}
}
\value{
junctions as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object with
additional \code{rowData} detailing overlapping genes/transcripts/exons and
junction categories.
}
\description{
\code{junction_annot} annotates junctions by 1. Whether their start and/or
end position precisely overlaps with an annotated exon boundary and 2. Whether
that junction matches any intron definition from existing annotation.
Using this information along with the strand, junctions are categorised into
"annotated", "novel_acceptor", "novel_donor", "novel_combo",
"novel_exon_skip", "ambig_gene" and "unannotated".
}
\examples{

load("data/junctions_example.rda")
junctions_example <-
    junction_annot(
        junctions_example,
        "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
    )
junctions_example

}
