% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_annot.R
\name{junction_annot}
\alias{junction_annot}
\title{Annotate junctions using reference annotation}
\usage{
junction_annot(junctions, ref)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{ref}{either path to gtf/gff3 or object of class \code{TxDb} imported using
\code{\link[GenomicFeatures]{makeTxDbFromGFF}}.}
}
\value{
junctions as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object with additional \code{rowData} detailing overlapping
genes/transcripts/exons and junction categories.
}
\description{
\code{junction_annot} annotates junctions by 1. Whether their start and/or
end position precisely overlaps with an annotated exon boundary and 2.
Whether that junction matches any intron definition from existing annotation.
Using this information along with the strand, junctions are categorised into
"annotated", "novel_acceptor", "novel_donor", "novel_combo",
"novel_exon_skip", "ambig_gene" and "unannotated".
}
\examples{

ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"
if (!exists("junctions_annoted")) {
    junctions_annoted <-
        junction_annot(
            junctions_example,
            ref
        )
}
junctions_annoted
}
