% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_filter.R
\name{junction_filter}
\alias{junction_filter}
\title{Filter junctions}
\usage{
junction_filter(
  junctions,
  count_thresh = c(raw = 5),
  n_samp = c(raw = 1),
  width_range = NULL,
  types = NULL,
  regions = NULL
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{count_thresh}{named vector with names matching the names of the
\code{\link[SummarizedExperiment]{assays}} in \code{junctions}. Values
denote the number of counts below which a junction will be filtered out.}

\item{n_samp}{named vector with names matching the names of the
\code{\link[SummarizedExperiment]{assays}} in \code{junctions}. Values
denotes number of samples that have to express the junction above the
\code{count_thresh} in order for that junction to not be filtered.}

\item{width_range}{numeric vector of length 2. The first element denoting the
lower limit of junction width and the second the upper limit. Junctions
with widths outside this range will be filtered out.}

\item{types}{any junctions matching these types, derived form
\code{\link{junction_annot}} e.g. "unannotated" will be filtered out.}

\item{regions}{any junctions overlapping this set of regions (in a
\code{\link[GenomicRanges]{GRanges}} format) will be filtered out.}
}
\value{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
containing filtered set of junctions.
}
\description{
\code{junction_filter} filters out "noisy" junctions based on counts, the
width of junctions, annotation category of the junction returned from
\code{\link{junction_annot}} and whether the junction overlaps with a set of
(blacklist) regions.
}
\examples{

junctions_filtered <- junction_filter(junctions_example)
junctions_filtered
}
\seealso{
ENCODE blacklist regions recommended to  be included as
\code{regions} can be downloaded from
\url{https://github.com/Boyle-Lab/Blacklist/blob/master/lists/hg38-blacklist.v2.bed.gz}.
Further information can be found via the publication
\url{https://www.nature.com/articles/s41598-019-45839-z}.
}
