% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_norm.R
\name{junction_norm}
\alias{junction_norm}
\title{Normalise junction counts}
\usage{
junction_norm(junctions)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}
}
\value{
junctions as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object with an
additional \code{assay} containing the normalised counts.
}
\description{
\code{junction_norm} normalises the junction counts for each sample. First,
each junction is clustered by finding all other junctions that share with it
an acceptor or donor position. Then, a proportion-spliced-in (PSI) is
calculated for each junction by dividing the raw junction count by the total
number of counts in it's associated cluster.
}
\examples{

junctions <- junction_norm(junctions_example)
junctions
}
