% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_norm.R
\name{junction_norm}
\alias{junction_norm}
\title{Normalise junction counts by cluster}
\usage{
junction_norm(junctions)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}
}
\value{
junctions as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object with an additional `assay`` containing the normalised counts.
}
\description{
\code{junction_norm} normalises the raw junction counts by 1. building junction
clusters by finding junctions that share an acceptor or donor position and 2.
calculating a proportion-spliced-in (PSI) for each junction by dividing the
raw junction count by the total number of counts in it's associated cluster.
}
\examples{

if (!exists("junctions_normed")) {
    junctions_normed <- junction_norm(junctions_example)
}
junctions_normed
}
\seealso{
Junction processing functions
\code{\link{junction_annot}()},
\code{\link{junction_filter}()},
\code{\link{junction_load}()},
\code{\link{junction_process}()},
\code{\link{junction_score}()}
}
\concept{junction}
