% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_process.R
\name{junction_process}
\alias{junction_process}
\title{Wrapper for processing junctions}
\usage{
junction_process(
  junctions,
  ref,
  count_thresh = c(raw = 5),
  n_samp = c(raw = 1),
  width_range = NULL,
  types = NULL,
  regions = NULL,
  score_func = .zscore,
  ...
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{ref}{either path to gtf/gff3 or object of class \link[GenomicFeatures:TxDb-class]{TxDb-class}.}

\item{count_thresh}{named vector with names matching the names of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assays} in \code{junctions}. Values
denote the number of counts below which a junction will be filtered out.}

\item{n_samp}{named vector with names matching the names of the
\link[SummarizedExperiment:SummarizedExperiment-class]{assays} in \code{junctions}. Values
denotes number of samples that have to express the junction above the
\code{count_thresh} in order for that junction to not be filtered.}

\item{width_range}{numeric vector of length 2. The first element denoting the
lower limit of junction width and the second the upper limit. Junctions
with widths outside this range will be filtered out.}

\item{types}{any junctions matching these types, derived form
\link{junction_annot} will be filtered out.}

\item{regions}{any junctions overlapping this set of regions (in a
\link[GenomicRanges:GRanges-class]{GRanges-class} format) will be filtered out.}

\item{score_func}{function to score junctions by their abnormality. By default,
will use a z-score but can be switched to a user-defined function. This
function must take as input an \code{x} and \code{y} argument, containing case and
control counts respectively. This must return a numeric vector equal to the
length of \code{x} with elements corresponding to a abnormality of each junction.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object containing filtered, annotated, normalised junction data with abnormality scores.
}
\description{
\code{junction_process} wraps all "junction_" prefixed functions in
\link[=dasper]{dasper} except \link{junction_load}. This is designed to simplify
processing of the junction data for those familiar or uninterested with the
intermediates.
}
\examples{

ref <-
    "ftp://ftp.ensembl.org/pub/release-100/gtf/homo_sapiens/Homo_sapiens.GRCh38.100.gtf.gz"

if (!exists("junctions_processed")) {
    junctions_processed <-
        junction_process(
            junctions_example,
            ref,
            count_thresh = c("raw" = 5),
            n_samp = c("raw" = 1),
            width_range = c(25, 1000000),
            types = c("ambig_gene", "unannotated"),
        )
}
junctions_processed
}
\seealso{
Junction processing functions
\code{\link{junction_annot}()},
\code{\link{junction_filter}()},
\code{\link{junction_load}()},
\code{\link{junction_norm}()},
\code{\link{junction_score}()}
}
\concept{junction}
