% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_score.R
\name{junction_score}
\alias{junction_score}
\title{Score patient junctions in terms of abnormality in relation to controls}
\usage{
junction_score(junctions, score_func = .zscore, ...)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{score_func}{function to score junctions by abnromality. By default,
will use a z-score. But this can be switched dependent on the format of the
user's junction data. Function must take as input an x and y argument. x
being the case junctions and y being the controls. This must return a
numeric vector equal to the length of x with each element corresponding to
a abnormality of a junction in each case sample.}

\item{...}{additional arguments passed to \code{score_func}.}
}
\value{
junctions as a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object filtered
for only "case" samples with an additional \code{assay} containing junction
abnormality scores.
}
\description{
\code{junction_score} will use the counts contained within the "norm"
\code{\link[SummarizedExperiment]{assay}} to calculate a deviation of each
patient junction from the expected distribution of controls counts. The
function used to calculate this abnormality score can be user-inputted or
left as the default z-score.
}
\examples{

junctions <- junction_norm(junctions_example)
junctions <- junction_score(junctions)
junctions
}
