% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_junc.R
\name{merge_junc}
\alias{merge_junc}
\title{Merge junction data from multiple samples}
\usage{
merge_junc(
  junc_paths,
  sample_ids,
  load_func = .load_STAR,
  chr_to_filter = NULL
)
}
\arguments{
\item{junc_paths}{paths to files containing junction data.}

\item{sample_ids}{sample ids in the same order as \code{junc_paths}.}

\item{load_func}{function to load in junctions. By default, requires STAR
formatted junctions. But this could be switched dependent on user-specific
junction data format.}

\item{chr_to_filter}{chromosomes you would like to keep. By default, no
filter is applied.}
}
\value{
junction co-ordinates and the raw counts from each sample.
}
\description{
\code{merge_junc} merges the junction data from several samples. Uses a
\code{\link[dplyr]{full_join}}, so retains all unique junctions from every
sample.
}
