% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detect.R
\name{outlier_detect}
\alias{outlier_detect}
\title{Detecting outlier junctions}
\usage{
outlier_detect(junctions, feature_names = c("score", "coverage_score"), ...)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{feature_names}{names of assays in \code{junctions} that are to be used
as input into the outlier detection model.}

\item{...}{additional arguments passed to the outlier detection model
(isolation forest) for setting parameters.}
}
\value{
junctions as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object with additional assays named \code{outlier_scores}. The lower the
outlier score, the more of an outlier a junction is observed to be.
}
\description{
\code{outlier_detect} will use the features in
\code{\link[SummarizedExperiment]{assays}} named \code{feature_names} as
input into an unsupervised outlier detection algorithm to score each junction
based on how outlier-y it looks in relation to other junctions in the
patient. The default \code{score} and \coverage{outlier_scores} can be
calculated using the \code{\link{junction_process}} and
\code{\link{coverage_process}} respectively.
}
\seealso{
for more details on the isolation forest model used:
https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.IsolationForest.html
}
