% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_process.R
\name{outlier_process}
\alias{outlier_process}
\title{Wrapper for processing outliers}
\usage{
outlier_process(
  junctions,
  feature_names = c("score", "coverage_score"),
  samp_id_col = "samp_id",
  bp_param = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{junctions}{junction data as a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.}

\item{feature_names}{names of assays in \code{junctions} that are to be used as
input into the outlier detection model.}

\item{samp_id_col}{name of the column in the
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}
that details the sample ids.}

\item{bp_param}{a
\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam-class} instance
denoting whether to parallelise the calculating of outlier scores across
samples.}

\item{...}{additional arguments passed to the outlier detection model
(isolation forest) for setting parameters.}
}
\value{
\code{DataFrame} with one row per cluster detailing each cluster's
associated junctions, outlier scores, ranks and genes.
}
\description{
\code{outlier_process} wraps all "outlier_" prefixed functions in
\link{dasper}. This is designed to simplify processing of the detecting
outlier junctions for those familiar or uninterested with the intermediates.
}
\seealso{
Outlier processing functions
\code{\link{outlier_aggregate}()},
\code{\link{outlier_detect}()}
}
\concept{outlier}
