% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ref_load}
\alias{ref_load}
\title{Load reference annotation into a TxDb format}
\usage{
ref_load(ref)
}
\arguments{
\item{ref}{either path to gtf/gff3 or object of class
\link[GenomicFeatures:TxDb-class]{TxDb-class} or
\link[ensembldb:EnsDb-class]{EnsDb-class}.
\link[ensembldb:EnsDb-class]{EnsDb-class} is required if you intend to annotate
junctions with gene symbols/names.}
}
\value{
a \link[GenomicFeatures:TxDb-class]{TxDb-class} object.
}
\description{
\verb{ref_load`` loads reference annotation using [makeTxDbFromGFF][GenomicFeatures::makeTxDbFromGFF] if a character or leaves }ref` unchanged if already a \link[GenomicFeatures:TxDb-class]{TxDb-class}. If
you would
}
\examples{

# create a TxDb,
ref <- GenomicState::GenomicStateHub(
    version = "31",
    genome = "hg38",
    filetype = "TxDb"
)[[1]]

# alternatively ref can be a character specifying a path to a GTF file
ref <- ref_load(ref)
}
